login --> verification, removal of logout / login (from user settings)
